+++ main.py
@@ -1,0 +1,14 @@
+from PyQt5.QtWidgets import QApplication
+import sys
+from pde_solver import MainWindow
+
+
+def main():
+    app = QApplication(sys.argv)
+    window = MainWindow()
+    window.show()
+    sys.exit(app.exec_())
+
+
+if __name__ == "__main__":
+    main()
