+++ pde_solver/__init__.py
@@ -1,0 +1,18 @@
+from .gui import MainWindow, GLWidget, ControlPanel, ColorScaleWidget
+from .solvers import (
+    heat_equation_update,
+    schrodinger_equation_update,
+    schrodinger_equation_cn_update,
+    wave_equation_update,
+    burgers_equation_update,
+    navier_stokes_update,
+    reaction_diffusion_update,
+)
+
+__all__ = [
+    'MainWindow', 'GLWidget', 'ControlPanel', 'ColorScaleWidget',
+    'heat_equation_update', 'schrodinger_equation_update',
+    'schrodinger_equation_cn_update', 'wave_equation_update',
+    'burgers_equation_update', 'navier_stokes_update',
+    'reaction_diffusion_update'
+]
